requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
commonmark==0.9.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
djangorestframework==3.14.0
filelock==3.9.0
httpie==3.2.1
idna==3.4
multidict==6.0.4
platformdirs==2.6.2
protobuf==4.21.12
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.1.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.17.1
